fix: added selected tab in session storage and refactored code
